feature[test]: refactoring issue #95
